requirements.txt
flask
mysql-connector==2.1.4
requests
boto
bcrypt